+++ assets/javascripts/kitten/index.js
@@ -146,6 +146,7 @@ export { LinkList } from './components/links/link-list'
 export { IconList } from './components/lists/icon-list'
 export { BulletList } from './components/lists/bullet-list'
 export { List } from './components/lists/list'
+export { CheckList } from './components/lists/check-list'
 export { NavList } from './components/lists/nav-list'
 export { TagList } from './components/lists/tag-list'
 export { Timeline } from './components/lists/timeline'
@@ -236,3 +237,4 @@ export { Uploader } from './components/uploaders/uploader'
 
 // Videos
 export { Video } from './components/videos/video'
+export { GifVideo } from './components/videos/gif-video'